+++ src/main/java/racingcar/View/InputView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.View;
 
 public class InputView {
     private final String CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
